Re-work initial text line generation logic
To begin with (first pass) the first minute is always the last
whole minute (which can be up one minute in the past). As the
loop progresses, markerAngle is 'increasing' so we begin
generating text lines projected into the future. We continue
in this vain through the 'round the back' zone. Once we emerge
out of this invisible rear quadrant we switch to generating
historical text lines, since they may still be visible before
disappearing over the top of the cylinder.