Restore _loadTextures()
Since Gemini-CLI modified _loadTextures() to use ImageBitmap we lost
the asynchronous nature. It should (i) use Promise.all to fetch all the
texture assets in parallel then (ii) return a Promise so the hole thing
can be used as an async function (e.g. if the caller wanted to load
textures and fonts in parallel).